Move Tool & License Views tests to their own file
